third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger/debugger-scripts-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that valid parsed script notifications are received by front-end.
script 1:
    start: 1:12
    end: 5:0
script 2:
    start: 6:11
    end: 6:32
script 3:
    start: 7:11
    end: 8:0
script 4:
    start: 9:11
    end: 11:0
